moved structure_2d_list_form from over sample to maker
