+++ autoarray/operators/over_sample/abstract.py
@@ -22,23 +22,4 @@ class AbstractOverSampleFunc:
     def tree_unflatten(cls, aux_data, children):
         return cls(mask=children[0])
 
-    def structure_2d_list_from(
-        self,
-        result_list: List,
-    ) -> List[Union[Array2D, "Grid2D"]]:
-        """
-        Convert a result from a list of ndarrays to a list of aa.Array2D or aa.Grid2D structure, where the conversion
-        depends on type(result) as follows:
 
-        - [1D np.ndarray] -> [aa.Array2D]
-        - [2D np.ndarray] -> [aa.Grid2D]
-
-        This function is used by the grid_like_list_to_structure-list decorator to convert the output result of a
-        function to a list of autoarray structure when a `Grid2D` instance is passed to the decorated function.
-
-        Parameters
-        ----------
-        result_list or [np.ndarray]
-            The input result (e.g. of a decorated function) that is converted to a PyAutoArray structure.
-        """
-        return [self.structure_2d_from(result=result) for result in result_list]
